requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
audioread==2.1.9
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cycler==0.10.0
Cython==0.29.23
decorator==5.1.0
dill==0.3.4
Distance==0.1.3
einops==0.3.0
ffmpeg==1.4
Flask==2.0.1
Flask-API==3.0.post1
fsspec==2021.8.1
future==0.18.2
g2p-en==2.1.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.40.0
idna==3.2
inflect==5.0.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
jsonpatch==1.32
jsonpointer==2.1
kiwisolver==1.3.2
librosa==0.8.0
llvmlite==0.37.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.3
multidict==5.1.0
multiprocess==0.70.12.2
nltk==3.6.2
numba==0.54.0
numpy==1.19.5
noisereduce==2.0.0
oauthlib==3.1.1
omegaconf==2.0.6
opencv-python==4.5.3.56
packaging==21.0
pathlib==1.0.1
Pillow==8.3.2
pooch==1.5.1
prefetch-generator==1.0.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pynndescent==0.5.4
pyparsing==2.4.7
pypinyin==0.39.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
pytorch-lightning==1.2.10
PyYAML>=5.1
pyzmq==22.2.1
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
resampy==0.2.2
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
sounddevice==0.4.2
SoundFile==0.10.3.post1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tgt==1.4.4
threadpoolctl==2.2.0
torchfile==0.1.0
torchmetrics==0.2.0
tornado==6.1
tqdm==4.62.2
typing-extensions==3.10.0.2
umap-learn==0.5.1
Unidecode==1.1.2
urllib3==1.26.6
visdom==0.1.8.9
webrtcvad==2.0.10
websocket-client==1.2.1
Werkzeug==2.0.1
yarl==1.6.3